fix: use the port of the annotation when specified
When the `metricsPortNumber` annotation value is set, always use that port for the scraping, ignoring the declared container ports of the pod.

This is the [current Prometheus approach](https://github.com/prometheus-community/helm-charts/blob/cd640243eb03787fab11cf2883534f604787f760/charts/prometheus/values.yaml#L1068-L1077) and ensures the pod is scraped even if the container port is not declared.

For example, this overcomes an [issue](https://github.com/istio/istio/issues/47882) where pods with istio-proxy are not scraped because Istio only declares the Envoy Prometheus telemetry port (15090) on the istio-proxy container and not also the [Merged Prometheus telemetry port](https://istio.io/latest/docs/ops/integrations/prometheus/#option-1-metrics-merging) (15020), which is instead correctly set as `prometheus.io/port` annotation on the pod.